Consistency with other SDKs
