+++ transaction.go
@@ -26,40 +26,6 @@ const (
 	TransactionStatusUnrecognized           TransactionStatus = "unrecognized"
 )
 
-const (
-	ThreeDSecureStatusUnsupportedCard                      = "unsupported_card"
-	ThreeDSecureStatusLookupError                          = "lookup_error"
-	ThreeDSecureStatusLookupEnrolled                       = "lookup_enrolled"
-	ThreeDSecureStatusLookupNotEnrolled                    = "lookup_not_enrolled"
-	ThreeDSecureStatusAuthSuccessfulIssuerNotParticipating = "authenticate_successful_issuer_not_participating"
-	ThreeDSecureStatusAuthUnavailable                      = "authentication_unavailable"
-	ThreeDSecureStatusAuthSignatureVerificationFailed      = "authenticate_signature_verification_failed"
-	ThreeDSecureStatusAuthSuccessful                       = "authenticate_successful"
-	ThreeDSecureStatusAuthAttemptSuccessful                = "authenticate_attempt_successful"
-	ThreeDSecureStatusAuthFailed                           = "authenticate_failed"
-	ThreeDSecureStatusAuthUnableToAuthenticate             = "authenticate_unable_to_authenticate"
-	ThreeDSecureStatusAuthError                            = "authenticate_error"
-)
-
-// Enrollment values can be found here:
-// https://developers.braintreepayments.com/reference/response/transaction/php#three_d_secure_info.enrolled
-const (
-	ThreeDSecureEnrollementYes            = "Y"
-	ThreeDSecureEnrollementNo             = "N"
-	ThreeDSecureEnrollementUnavailable    = "U"
-	ThreeDSecureEnrollementBypass         = "B"
-	ThreeDSecureEnrollementRequestFailure = "E"
-)
-
-// RiskData possible values
-// https://developers.braintreepayments.com/guides/advanced-fraud-tools/server-side/java#response-handling
-const (
-	RiskDataNotEvaluated = "Not Evaluated"
-	RiskDataApprove      = "Approve"
-	RiskDataReview       = "Review"
-	RiskDataDecline      = "Decline"
-)
-
 type TransactionSource string
 
 const (
@@ -221,10 +187,6 @@ type Transactions struct {
 	Transaction []*Transaction `xml:"transaction"`
 }
 
-type TransactionOptionsThreeDSecure struct {
-	Required bool `xml:"required"`
-}
-
 type TransactionOptions struct {
 	SubmitForSettlement              bool                             `xml:"submit-for-settlement,omitempty"`
 	StoreInVault                     bool                             `xml:"store-in-vault,omitempty"`
@@ -292,6 +254,10 @@ func (r TransactionOptionsPaypalRequest) MarshalXML(e *xml.Encoder, start xml.St
 	return nil
 }
 
+type TransactionOptionsThreeDSecure struct {
+	Required bool `xml:"required"`
+}
+
 type TransactionSearchResult struct {
 	TotalItems int
 	TotalIDs   []string
@@ -309,13 +275,6 @@ type RiskData struct {
 type RiskDataRequest struct {
 	CustomerBrowser string `xml:"customer-browser"`
 	CustomerIP      string `xml:"customer-ip"`
-}
-
-type ThreeDSecureInfo struct {
-	Enrolled               string `xml:"enrolled"`
-	LiabilityShiftPossible bool   `xml:"liability-shift-possible"`
-	LiabilityShifted       bool   `xml:"liability-shifted"`
-	Status                 string `xml:"status"`
 }
 
 type SubscriptionDetails struct {
